Rename downloadWithConcurrencyLimit to downloadTracks/index.ts
