+++ src/lib/index.ts
@@ -2,8 +2,8 @@ import getBody from './getBody';
 import getTracks from './getTracks';
 import validateTracks from './validateTracks';
 import logger from './logger';
-import downloadWithConcurrencyLimit from './downloadTracks/downloadWithConcurrencyLimit';
+import downloadTracks from './downloadTracks';
 
 export {
-  getBody, getTracks, validateTracks, logger, downloadWithConcurrencyLimit,
+  getBody, getTracks, validateTracks, logger, downloadTracks,
 };
